Strangler extraction #14: hdfmonkey install chain -> zxnu_emulator_ops.py
Move the hdfmonkey download/install/discovery chain (the jjjs
zip-inside-a-zip download + per-platform extract, the manual-download
fallback prompt, the install-button flow, _hdfmonkey_binary_found and
the missing-hdfmonkey prompt signal wiring, ~390 lines) out of
MainWindow.__init__ into a second builder in zxnu_emulator_ops.py,
build_hdfmonkey_install_ops(host, ...).

Same seam as the rest: definition-only block, self -> host token pass,
button-animation/load_image/log helpers injected as forwarding lambdas,
the three closures the SD-card wiring consumes by bare name exposed on
host and re-bound at the call site. os.path.dirname(__file__) inside
_hdfmonkey_binary_found still resolves to the app directory (all
modules ship side by side in source, PyInstaller and pip layouts).
zxnu_main.py 8,580 -> 8,204 lines.

test_pane_imports.py: treat module dunders (__file__/__name__) as bound
in the free-var tripwire - they always resolve at runtime.
