+++ zxnu_emulator_ops.py
@@ -1575,3 +1575,424 @@ def build_emulator_ops(
     host.set_mame_esc = set_mame_esc
     host.open_cspect_configuration_file = open_cspect_configuration_file
     host.install_mame = install_mame
+
+
+def build_hdfmonkey_install_ops(
+    host,
+    *,
+    _update_mame_controls,
+    _update_cspect_controls,
+    _start_hdfmonkey_button_animation,
+    _stop_hdfmonkey_button_animation,
+    load_image,
+    add_main_log_window,
+):
+    """Define the hdfmonkey download/install/discovery chain (no widgets).
+
+    The jjjs zip-inside-a-zip download + per-platform extract, the manual-
+    download fallback prompt, the install-button flow, binary discovery
+    (_hdfmonkey_binary_found — note os.path.dirname(__file__) still resolves
+    to the app directory: every module ships side by side in source,
+    PyInstaller and pip layouts alike) and the missing-hdfmonkey prompt
+    signal wiring.
+    """
+    def _hdfmonkey_downloads_root():
+        """Top-level ``downloads`` folder next to the app (created if needed).
+        This is where the auto-download saves the jjjs zip and where the
+        manual-fallback flow asks the user to drop a hand-downloaded copy."""
+        app_dir = ZXNU_DATA_ROOT
+        root = os.path.join(app_dir, DOWNLOADS_ROOT_DIRNAME)
+        os.makedirs(root, exist_ok=True)
+        return root
+
+    def _download_hdfmonkey_zip():
+        """STEP 1 — download the jjjs hdfmonkey archive from
+        HDF_MONKEY_JJJS_URL into the downloads folder.
+
+        Logs each stage (URL, HTTP status, content type/length, bytes read,
+        save location) so a failure points at the real cause. Returns the
+        saved zip path on success, or None on failure — importantly it detects
+        the common specnext.com case where an HTML login / anti-robot page is
+        returned with a 200 status instead of the actual attachment, rather
+        than letting that non-zip flow through to a confusing extract error.
+        """
+        downloads_root = _hdfmonkey_downloads_root()
+        dest_zip = os.path.join(downloads_root, HDF_MONKEY_JJJS_ZIP_FILENAME)
+        add_main_log_window(f"hdfmonkey: [1/2] downloading archive from {HDF_MONKEY_JJJS_URL} ...")
+        logging.info(f"hdfmonkey download: requesting {HDF_MONKEY_JJJS_URL}")
+        try:
+            req = urllib.request.Request(
+                HDF_MONKEY_JJJS_URL, headers={"User-Agent": ZXART_USER_AGENT})
+            with urllib.request.urlopen(req, timeout=60) as resp:
+                status = getattr(resp, "status", None) or resp.getcode()
+                content_type = resp.headers.get("Content-Type", "") or ""
+                content_length = resp.headers.get("Content-Length", "") or ""
+                final_url = resp.geturl()
+                add_main_log_window(
+                    f"hdfmonkey: server responded HTTP {status}; "
+                    f"Content-Type='{content_type}', "
+                    f"Content-Length='{content_length or 'unknown'}'.")
+                logging.info(
+                    f"hdfmonkey download: HTTP {status} type={content_type!r} "
+                    f"len={content_length!r} final_url={final_url!r}")
+                data = resp.read()
+        except urllib.error.HTTPError as e:
+            add_main_log_window(
+                f"hdfmonkey: download failed — server returned HTTP {e.code} "
+                f"{e.reason}. specnext.com may require a forum login or an "
+                f"anti-robot confirmation before the file can be downloaded.")
+            logging.error(f"hdfmonkey download HTTPError: {e.code} {e.reason}")
+            return None
+        except urllib.error.URLError as e:
+            add_main_log_window(
+                f"hdfmonkey: download failed — could not reach "
+                f"{HDF_MONKEY_JJJS_URL} ({e.reason}). Check your internet "
+                f"connection, or a firewall/proxy blocking the request.")
+            logging.error(f"hdfmonkey download URLError: {e.reason}")
+            return None
+        except Exception as e:
+            add_main_log_window(f"hdfmonkey: download failed — {e}")
+            logging.error(f"hdfmonkey download error: {e}")
+            return None
+
+        add_main_log_window(f"hdfmonkey: downloaded {len(data):,} bytes.")
+        logging.info(f"hdfmonkey download: read {len(data)} bytes")
+
+        # Guard against the forum returning an HTML login / anti-robot / error
+        # page (often with a 200 status) instead of the real attachment. A
+        # genuine zip starts with the 'PK' local-file-header magic.
+        if data[:2] != b"PK":
+            head = data[:200].decode("utf-8", "replace").strip()
+            head = " ".join(head.split())
+            lower = data[:1024].lower()
+            looks_html = b"<html" in lower or b"<!doctype" in lower
+            if looks_html:
+                add_main_log_window(
+                    "hdfmonkey: the server returned a web page, not the zip "
+                    "file — this usually means specnext.com is asking for a "
+                    "login or an anti-robot confirmation before the download "
+                    "starts.")
+            else:
+                add_main_log_window(
+                    "hdfmonkey: the downloaded data is not a zip file "
+                    f"(it begins with: {head!r}).")
+            logging.error(
+                f"hdfmonkey download: not a zip (html={looks_html}); "
+                f"head={head!r}")
+            return None
+
+        try:
+            with open(dest_zip, "wb") as f:
+                f.write(data)
+        except OSError as e:
+            add_main_log_window(
+                f"hdfmonkey: could not save the downloaded zip to "
+                f"{dest_zip} — {e}")
+            logging.error(f"hdfmonkey download save error: {e}")
+            return None
+
+        # Verify the pinned SHA-256 BEFORE anything touches the archive:
+        # a corrupted or tampered download is refused, not extracted.
+        actual = sha256_of_file(dest_zip)
+        if actual.lower() != HDF_MONKEY_JJJS_SHA256:
+            add_main_log_window(
+                "hdfmonkey: SHA-256 mismatch on the downloaded archive "
+                f"(expected {HDF_MONKEY_JJJS_SHA256[:12]}…, got {actual[:12]}…) "
+                "— refusing to extract it. Retry the download, or install "
+                "hdfmonkey manually (see the wiki).")
+            logging.error(
+                f"hdfmonkey download hash mismatch: expected "
+                f"{HDF_MONKEY_JJJS_SHA256}, got {actual}")
+            try:
+                os.remove(dest_zip)
+            except OSError:
+                pass
+            return None
+        add_main_log_window(
+            f"hdfmonkey: saved archive to {dest_zip} (SHA-256 verified).")
+        logging.info(f"hdfmonkey download: saved {dest_zip}")
+        return dest_zip
+
+    def _install_hdfmonkey_from_zip(zip_path, keep_zip=False):
+        """STEP 2 — extract this platform's hdfmonkey binary out of a jjjs
+        archive (auto-downloaded or manually dropped) into
+        downloads/hdfmonkey/<platform>/.
+
+        The extracted binary is recorded so execute_hdf_monkey prefers it and
+        re-discovered on the next launch by find_hdfmonkey_in_downloads (so
+        this works on Windows, Linux and macOS). Returns the installed binary
+        path on success, or None on failure. ``keep_zip`` leaves a
+        user-provided archive in place; an auto-downloaded one is removed once
+        it has been unpacked.
+        """
+        app_dir = ZXNU_DATA_ROOT
+        dest_root = os.path.join(app_dir, DOWNLOADS_HDFMONKEY_DIRNAME)
+        add_main_log_window(
+            f"hdfmonkey: [2/2] extracting the build for this platform from "
+            f"{zip_path} ...")
+        logging.info(f"hdfmonkey install: extracting from {zip_path}")
+        try:
+            hdfmonkey_path = extract_hdfmonkey_from_jjjs_zip(zip_path, dest_root)
+        except Exception as e:
+            add_main_log_window(
+                f"hdfmonkey: could not extract the binary from {zip_path} — "
+                f"{e}. The archive may be incomplete or corrupted; try "
+                f"downloading it again.")
+            logging.error(f"hdfmonkey install extract error: {e}")
+            return None
+        finally:
+            if not keep_zip:
+                try:
+                    os.remove(zip_path)
+                except OSError:
+                    pass
+        add_main_log_window(f"hdfmonkey: extracted binary to {hdfmonkey_path}.")
+        logging.info(f"hdfmonkey install: extracted {hdfmonkey_path}")
+        return hdfmonkey_path
+
+    def _finish_hdfmonkey_install(hdfmonkey_path):
+        """Common UI updates once hdfmonkey has been installed (whether via
+        the automatic download or the manual-drop fallback)."""
+        host._hdfmonkey_executable_path = hdfmonkey_path
+        host.button_new_folder.setVisible(True)
+        host.button_rename.setVisible(True)
+        host.button_delete_files.setVisible(True)
+        host.download_and_install_hdfmonkey_button.setVisible(False)
+        # hdfmonkey is now installed — stop the yellow attention pulse and
+        # restore the button's normal look straight away.
+        _stop_hdfmonkey_button_animation()
+        logging.info(f"Successfully installed hdfmonkey: {hdfmonkey_path}")
+        add_main_log_window(f"Successfully installed hdfmonkey: {hdfmonkey_path}")
+
+        # Confirm the install with a green toast (like the emulator detection
+        # one) showing where the binary landed on disk.
+        host._show_hdfmonkey_installed_toast(hdfmonkey_path)
+
+        # Reload the currently-selected image straight away so the file
+        # explorer repopulates without the user having to reopen it via
+        # "Select NextZXOS disk Image". The extract succeeded and
+        # _hdfmonkey_executable_path now points at a verified binary, so
+        # there's no need to re-probe first; load_image() restores the
+        # controls once the (async) listing completes, and safely no-ops when
+        # no image is selected.
+        load_image()
+
+    def _try_install_hdfmonkey_from_manual_zip():
+        """Look for a jjjs hdfmonkey zip the user dropped into the downloads
+        folder and, if a valid one is found, install from it. Returns True on
+        a successful install."""
+        app_dir = ZXNU_DATA_ROOT
+        manual_zip = find_hdfmonkey_jjjs_zip_in_downloads(app_dir)
+        if not manual_zip:
+            return False
+        add_main_log_window(
+            f"hdfmonkey: found a manually-downloaded archive at "
+            f"{manual_zip} — installing from it.")
+        logging.info(f"hdfmonkey: installing from manual zip {manual_zip}")
+        # Leave a hand-placed archive in place so the user can retry offline.
+        hdfmonkey_path = _install_hdfmonkey_from_zip(manual_zip, keep_zip=True)
+        if hdfmonkey_path:
+            _finish_hdfmonkey_install(hdfmonkey_path)
+            return True
+        return False
+
+    def _prompt_manual_hdfmonkey_download():
+        """After an automatic download failure, invite the user to fetch the
+        zip in a browser and drop it into the downloads folder, then retry
+        detecting it from there. Returns True if a manually-dropped zip was
+        subsequently found and installed."""
+        downloads_root = _hdfmonkey_downloads_root()
+        box = QMessageBox(host)
+        box.setIcon(QMessageBox.Warning)
+        box.setWindowTitle("hdfmonkey download failed")
+        box.setText(
+            "The automatic hdfmonkey download from specnext.com failed — the "
+            "forum may be asking for a login or an anti-robot confirmation "
+            "before the download can start (see the log for details).\n\n"
+            "You can install it manually instead:\n"
+            f"1. Click 'Open download page' below (or browse to\n"
+            f"    {HDF_MONKEY_JJJS_URL} ).\n"
+            "2. Download the hdfmonkey .zip file.\n"
+            f"3. Drop the downloaded .zip into this folder:\n"
+            f"    {downloads_root}\n"
+            "4. Click \"I've dropped the zip - try again\".")
+        open_page_btn = box.addButton("Open download page", QMessageBox.ActionRole)
+        open_folder_btn = box.addButton("Open downloads folder", QMessageBox.ActionRole)
+        retry_btn = box.addButton("I've dropped the zip - try again", QMessageBox.AcceptRole)
+        box.addButton("Cancel", QMessageBox.RejectRole)
+        box.setDefaultButton(retry_btn)
+        while True:
+            box.exec()
+            clicked = box.clickedButton()
+            if clicked is open_page_btn:
+                try:
+                    webbrowser.open(HDF_MONKEY_JJJS_URL)
+                    add_main_log_window(
+                        f"hdfmonkey: opened {HDF_MONKEY_JJJS_URL} in your "
+                        f"browser. Save the .zip into {downloads_root}, then "
+                        f"click 'try again'.")
+                except Exception as e:
+                    add_main_log_window(
+                        f"hdfmonkey: could not open the browser automatically "
+                        f"({e}). Please browse to {HDF_MONKEY_JJJS_URL} "
+                        f"manually.")
+                continue  # reshow the dialog so the user can retry afterwards
+            if clicked is open_folder_btn:
+                try:
+                    if sys.platform == "win32":
+                        os.startfile(downloads_root)
+                    elif sys.platform == "darwin":
+                        subprocess.Popen(["open", downloads_root])
+                    else:
+                        subprocess.Popen(["xdg-open", downloads_root])
+                except Exception as e:
+                    add_main_log_window(
+                        f"hdfmonkey: could not open {downloads_root} ({e}).")
+                continue
+            if clicked is retry_btn:
+                if _try_install_hdfmonkey_from_manual_zip():
+                    return True
+                add_main_log_window(
+                    f"hdfmonkey: no valid hdfmonkey .zip found in "
+                    f"{downloads_root} yet. Download it from "
+                    f"{HDF_MONKEY_JJJS_URL} and drop the .zip there, then try "
+                    f"again.")
+                continue  # let the user place the file and retry
+            return False  # Cancel / dialog closed
+
+    def download_and_install_hdflonkey():
+        """Install hdfmonkey in two explicit, individually-logged steps:
+        download the jjjs archive, then extract this platform's binary. If a
+        valid archive is already sitting in the downloads folder (e.g. the
+        user dropped one after a previous failure) it is used directly; if the
+        automatic download fails, the manual browser-download fallback is
+        offered."""
+        # Use an already-present (manually-dropped) archive first — this makes
+        # a retry after the manual route succeed without hitting the network
+        # again, and keeps the whole flow working when specnext.com blocks us.
+        if _try_install_hdfmonkey_from_manual_zip():
+            return True
+
+        zip_path = _download_hdfmonkey_zip()
+        if not zip_path:
+            # Download blocked/failed — offer the manual browser route (drop
+            # the zip into downloads, then detect it from there).
+            return _prompt_manual_hdfmonkey_download()
+
+        hdfmonkey_path = _install_hdfmonkey_from_zip(zip_path, keep_zip=False)
+        if not hdfmonkey_path:
+            # A file downloaded but could not be unpacked — the cleanest
+            # recovery is a fresh (manual) download.
+            add_main_log_window(
+                "hdfmonkey: you can install it manually from "
+                "https://github.com/gasman/hdfmonkey , or (recommended) do a "
+                "full CSpect install from the itch.io tab, which also bundles "
+                "hdfmonkey.")
+            return _prompt_manual_hdfmonkey_download()
+
+        _finish_hdfmonkey_install(hdfmonkey_path)
+        return True
+
+    def _on_hdfmonkey_button_clicked():
+        """Button handler for "Download and install HDF Monkey". Shows an
+        intermediary tip about the itch.io end-to-end CSpect install (which
+        also bundles hdfmonkey) before running the standalone hdfmonkey
+        download, so the user can choose the fuller route instead."""
+        box = QMessageBox(host)
+        box.setIcon(QMessageBox.Information)
+        box.setWindowTitle("Install hdfmonkey")
+        box.setText(
+            "TIP: Did you know that if you have purchased CSpect from "
+            "itch.io you can do a full end-to-end CSpect install from "
+            "there?\n\n"
+            "Simply log into your itch.io account in the itch.io tab, "
+            "navigate to CSpect and click Install.\n\n"
+            "Do you still want to install hdfmonkey only, or abort and then "
+            "make an end-to-end install of CSpect using itch.io?")
+        continue_btn = box.addButton(
+            "Continue hdfmonkey standalone install", QMessageBox.AcceptRole)
+        box.addButton("Cancel", QMessageBox.RejectRole)
+        box.setDefaultButton(continue_btn)
+        box.exec()
+        if box.clickedButton() is continue_btn:
+            return download_and_install_hdflonkey()
+        return False
+
+    def show_hdf_monkey_download_and_install_buttons():
+        host.download_and_install_hdfmonkey_button.setVisible(True)
+        host.button_new_folder.setVisible(False)
+        host.button_rename.setVisible(False)
+        host.button_delete_files.setVisible(False)
+        # Draw the eye to the install button with a yellow 'breathing' pulse
+        # while hdfmonkey is missing; it stops itself once the button is
+        # hidden (i.e. hdfmonkey has been installed/detected).
+        _start_hdfmonkey_button_animation()
+        # hdfmonkey is confirmed missing here, so the file explorer stays
+        # disabled — but MAME doesn't need hdfmonkey, so make sure its launch
+        # button and option combos reflect (MAME present + a valid image)
+        # rather than staying stuck disabled. The CSpect option combos don't
+        # need hdfmonkey either (only Launch does, via the mounted image),
+        # so refresh them too when a CSpect build is present.
+        _update_mame_controls()
+        _update_cspect_controls()
+
+    def _hdfmonkey_binary_found():
+        """True if the hdfmonkey executable can be located (PATH, current
+        directory, the application directory, or a bundled copy discovered
+        under downloads/cspect). Used to tell a genuine hdfmonkey error apart
+        from "it isn't installed", without running it."""
+        override = getattr(host, "_hdfmonkey_executable_path", None)
+        if override and os.path.isfile(override):
+            return True
+        # A copy left by the standalone auto-download (downloads/hdfmonkey/
+        # <platform>/) counts too, even before startup re-adopts it as the
+        # active override (e.g. right after a fresh launch).
+        try:
+            if find_hdfmonkey_in_downloads(ZXNU_DATA_ROOT):
+                return True
+        except Exception:
+            pass
+        if shutil.which(HDFMONKEY_EXECUTABLE):
+            return True
+        names = [HDFMONKEY_EXECUTABLE]
+        if platform.system() == "Windows":
+            names.append(HDFMONKEY_EXECUTABLE + ".exe")
+        search_dirs = [os.getcwd(), ZXNU_DATA_ROOT]
+        try:
+            search_dirs.append(os.path.dirname(os.path.abspath(sys.argv[0])))
+        except Exception:
+            pass
+        try:
+            search_dirs.append(os.path.dirname(os.path.abspath(__file__)))
+        except Exception:
+            pass
+        for d in search_dirs:
+            for n in names:
+                if d and os.path.isfile(os.path.join(d, n)):
+                    return True
+        return False
+
+    def prompt_install_hdfmonkey():
+        """Offer to install hdfmonkey when it appears to be missing, on every
+        platform. The jjjs auto-download ships fixed builds for Windows, Linux
+        and macOS, so this runs the same "Install hdfmonkey" tip box + download
+        flow as the SD-card tab button (which also points the user at the
+        fuller end-to-end CSpect install via itch.io). Runs on the UI thread
+        (invoked via the missing-signal so it is safe from workers)."""
+        if _on_hdfmonkey_button_clicked():
+            # Installed OK — allow a fresh prompt if it ever breaks again.
+            host._hdfmonkey_prompt_shown = False
+
+    # Marshals the "hdfmonkey is missing" prompt onto the UI thread: the
+    # signal may be emitted from a worker thread (uploads/deletes), so the
+    # dialog must not be created inline there.
+    host._hdfmonkey_prompt_shown = False
+    host._hdfmonkey_missing_signals = HdfMonkeyMissingSignals()
+    host._hdfmonkey_missing_signals.missing.connect(prompt_install_hdfmonkey)
+
+    # Expose the closures the rest of __init__ wires by bare name (re-bound
+    # to bare locals at the call site).
+    host._hdfmonkey_binary_found = _hdfmonkey_binary_found
+    host.show_hdf_monkey_download_and_install_buttons = (
+        show_hdf_monkey_download_and_install_buttons)
+    host._on_hdfmonkey_button_clicked = _on_hdfmonkey_button_clicked
